moved some files, rewrote display menu css as proof of concept
