+++ .gitignore
@@ -1,3 +1,4 @@
 .vscode/
 package-lock.json
 package.json
+jsconfig.json
