+++ src/main.js
@@ -12,7 +12,7 @@ window.addEventListener("keydown", event => {
     let key = event.key.length > 1 ? event.key : event.key.toLowerCase(); // only characters are lowercase
     if (event.altKey) key = "Alt+" + key;
     if (event.ctrlKey) key = "Ctrl+" + key;
-    
+
     game.controls.onKeyDownRepeat(event, key);
     if (event.repeat) return;
     game.controls.onKeyDown(event, key);
