+++ src/mechanics/mechanics.js
@@ -1,7 +1,6 @@
 import { Game } from "../game.js";
 import { ClearLines } from "./clearlines.js";
 import { LockPiece } from "./locking.js";
-import pieces from "../data/pieces.json" with { type: "json" };
 
 export class Mechanics {
     board;
