+++ src/menus/modals.js
@@ -1,4 +1,5 @@
 import { Game } from "../game.js";
+import { GenerateMenus } from "./generate.js";
 
 export class ModalActions {
     open;
@@ -10,6 +11,8 @@ export class ModalActions {
     pblistStart = document.getElementById("PBlist");
     gamemodeStart = document.getElementById("startGamemodeList");
     statsStart = document.getElementById("startStatsList");
+    settingDialogs = [...document.getElementsByClassName("settingsBox")];
+    settings = [...document.getElementsByClassName("settingRow")];
 
     /**
      * @param {Game} game
@@ -18,6 +21,7 @@ export class ModalActions {
         this.game = game;
         this.actions = this.game.menuactions;
         game.menuactions.menus = this;
+        this.generate = new GenerateMenus(game);
     }
 
     openModal(id) {
@@ -25,7 +29,7 @@ export class ModalActions {
         if (id == "settingsPanel" && this.closing) return;
         if (id == "queueModify" && !this.game.settings.game.allowQueueModify) return;
         this.game.stopGameTimers()
-
+        console.log(id, this.getOptions(id))
         this.getOptions(id).forEach(setting => {
             let settingType = this.getSettingType(id);
             let newval;
@@ -45,18 +49,28 @@ export class ModalActions {
 
         document.getElementById(id).showModal();
 
-        if (id == "gameStatsDialog") this.displayStats();
-        if (id == "gamemodeDialog") this.highlightGamemodeInMenu();
-        if (id == "competitiveDialog") this.renderPBs();
+        if (id == "gameStatsDialog") this.generate.displayStats();
+        if (id == "gamemodeDialog") this.generate.highlightGamemodeInMenu();
+        if (id == "competitiveDialog") this.generate.renderPBs();
         if (id != "settingsPanel" && this.settingPanel.open) this.closeDialog(this.settingPanel);
         this.open = true;
         this.game.sounds.toggleSongMuffle(this.open);
+
+        // temp to init settings
+        if (id == "displayDialog") {
+            const box = document.getElementById("displayDialog").children[1];
+            const boxsettings = this.settings.filter(item => item.parentElement.parentElement.id == box.parentElement.id);
+            this.generate.updateSizes(box, boxsettings);
+        }
 
     }
 
     getOptions(id) {
         const options = [...document.getElementsByClassName("option")];
-        return options.filter(item => item.parentElement.parentElement.id == id)
+        const set = this.settings.map(i => i.children[1]);
+        const filt = options.filter(item => item.parentElement.parentElement.id == id)
+        const filt2 = set.filter(item => item.parentElement.parentElement.parentElement.id == id)
+        return [...filt, ...filt2];
     }
 
     getSettingType(id) {
@@ -115,94 +129,6 @@ export class ModalActions {
         this.game.sounds.toggleSongMuffle(this.open);
         element.classList.add("closingAnimation");
         element.addEventListener("animationend", closingAnimation);
-    }
-
-    // GENERATE MENUS
-    generateGamemodeMenu() {
-        this.game.modes.getGamemodeNames().forEach(name => {
-            const setting = this.game.modes.getGamemodeJSON(name);
-            const button = document.createElement("button");
-            button.id = name;
-            button.classList = "gamemodeSelect";
-            button.textContent = setting.displayName;
-            button.addEventListener("click", () => {
-                menu.setGamemode(name);
-                modal.closeModal("gamemodeDialog");
-            });
-            this.gamemodeStart.parentNode.insertBefore(button, this.gamemodeStart);
-        })
-        this.gamemodeStart.remove();
-    }
-
-    highlightGamemodeInMenu() {
-        const gamemodeSelect = [...document.getElementsByClassName("gamemodeSelect")];
-        gamemodeSelect.forEach(setting => {
-            setting.classList.remove("selected");
-            if (setting.id == this.game.settings.game.gamemode)
-                setting.classList.add("selected");
-        });
-    }
-
-    renderPBs() {
-        const previous = [...document.getElementsByClassName("pbbox")];
-        previous.forEach(el => el.remove());
-
-        const pbs = this.game.profilestats.personalBests;
-        Object.keys(pbs).forEach(mode => {
-            const score = pbs[mode].score
-            const pbbox = document.createElement("div");
-
-            const text1 = document.createElement("h2")
-            text1.textContent = mode[0].toUpperCase() + mode.slice(1) + ': ';
-            const text2 = document.createElement("h2")
-            text2.textContent = score + this.game.modes.getSuffix(mode);
-            const clearbutton = document.createElement("button");
-            clearbutton.textContent = "X";
-            clearbutton.addEventListener("click", (event) => {
-                event.stopPropagation();
-                this.game.profilestats.removePB(mode);
-                pbbox.remove()
-            });
-            pbbox.appendChild(text1);
-            pbbox.appendChild(text2);
-            pbbox.appendChild(clearbutton);
-            pbbox.addEventListener("click", () => {
-                let el = document.createElement("a");
-                el.setAttribute("href", "data:text/plain;charset=utf-8," + encodeURIComponent(JSON.stringify(pbs[mode])));
-                el.setAttribute("download", `${mode}_pb.json`);
-                document.body.appendChild(el);
-                el.click();
-                document.body.removeChild(el);
-            })
-
-            pbbox.classList = "pbbox settingLayout";
-            this.pblistStart.parentNode.insertBefore(pbbox, this.pblistStart);
-        })
-    }
-
-    displayStats() {
-        const previous = [...document.getElementsByClassName("statText")];
-        previous.forEach(el => el.remove());
-
-        const stats = Object.getOwnPropertyNames(this.game.stats);
-        const skip = ['clearCols', 'clearPieces', 'game']
-        stats.forEach(stat => {
-            if (skip.includes(stat)) return;
-            let score = this.game.stats[stat]
-            if (stat == "tspins") score = score.reduce((a, b) => a + b, 0)
-            const statItem = document.createElement("p");
-            statItem.classList = "statText";
-
-            const text1 = document.createElement("span")
-            text1.classList = "spanright"
-            text1.textContent = stat + ":"
-            const text2 = document.createElement("span")
-            text2.classList = "spanleft"
-            text2.textContent = Math.round(score * 1000) / 1000
-            statItem.appendChild(text1);
-            statItem.appendChild(text2);
-            this.statsStart.parentNode.insertBefore(statItem, this.statsStart);
-        })
     }
 }
