+++ styles/menus.css
@@ -281,7 +281,6 @@ button>img {
 }
 
 .textInput {
-    width: 50%;
     height: 2.4vh;
     right: 5%;
     padding: 0.5%;
